more setup for textoverlay system
